cleaned up play
